moved button group to separate component
